refactor(attr-attribute): move into models namespace
non-breaking change since all models already require the models namespace for Identifiable
